[MRRTF-209] Migrate MCH precluster I/O to IO and DevIO libs
