+++ Detectors/MUON/MCH/IO/src/preclusters-reader-workflow.cxx
@@ -11,21 +11,23 @@
 
 #include <vector>
 #include "Framework/ConfigParamSpec.h"
-#include "MCHWorkflow/PreClusterWriterSpec.h"
-#include "Framework/CompletionPolicyHelpers.h"
+#include "MCHIO/PreClusterReaderSpec.h"
+#include "DetectorsRaw/HBFUtilsInitializer.h"
+#include "Framework/CallbacksPolicy.h"
 
 using namespace o2::framework;
 
-void customize(std::vector<o2::framework::CompletionPolicy>& policies)
+// ------------------------------------------------------------------
+void customize(std::vector<o2::framework::CallbacksPolicy>& policies)
 {
-  // ordered policies for the writers
-  policies.push_back(CompletionPolicyHelpers::consumeWhenAllOrdered(".*(?:MCH|mch).*[W,w]riter.*"));
+  o2::raw::HBFUtilsInitializer::addNewTimeSliceCallback(policies);
 }
 
 // we need to add workflow options before including Framework/runDataProcessing
 void customize(std::vector<ConfigParamSpec>& workflowOptions)
 {
   workflowOptions.emplace_back("enable-mc", VariantType::Bool, false, ConfigParamSpec::HelpString{"Propagate MC info"});
+  o2::raw::HBFUtilsInitializer::addConfigOption(workflowOptions);
 }
 
 #include "Framework/runDataProcessing.h"
@@ -33,5 +35,7 @@ void customize(std::vector<ConfigParamSpec>& workflowOptions)
 WorkflowSpec defineDataProcessing(const ConfigContext& config)
 {
   bool useMC = config.options().get<bool>("enable-mc");
-  return WorkflowSpec{o2::mch::getPreClusterWriterSpec(useMC, "mch-precluster-writer")};
+  WorkflowSpec wf{o2::mch::getPreClusterReaderSpec(useMC, "mch-precluster-reader")};
+  o2::raw::HBFUtilsInitializer hbfIni(config, wf);
+  return wf;
 }
