+++ Detectors/MUON/MCH/Workflow/CMakeLists.txt
@@ -20,8 +20,6 @@ o2_add_library(MCHWorkflow
                    src/DataDecoderSpec.cxx
                    src/ErrorReaderSpec.cxx
                    src/ErrorWriterSpec.cxx
-                   src/PreClusterReaderSpec.cxx
-                   src/PreClusterWriterSpec.cxx
                    src/TrackReaderSpec.cxx
                    src/TrackTreeReader.cxx
                    src/TrackWriterSpec.cxx
@@ -75,24 +73,6 @@ o2_add_executable(
         SOURCES src/EventFinderSpec.cxx src/event-finder-workflow.cxx
         COMPONENT_NAME mch
         PUBLIC_LINK_LIBRARIES O2::Framework O2::MCHTriggering)
-
-o2_add_executable(
-        preclusters-sink-workflow
-        SOURCES src/PreClusterSinkSpec.cxx src/preclusters-sink-workflow.cxx
-        COMPONENT_NAME mch
-        PUBLIC_LINK_LIBRARIES O2::Framework O2::DataFormatsMCH O2::MCHBase O2::MCHMappingImpl3)
-
-o2_add_executable(
-        preclusters-writer-workflow
-        SOURCES src/preclusters-writer-workflow.cxx
-        COMPONENT_NAME mch
-        PUBLIC_LINK_LIBRARIES O2::MCHWorkflow)
-
-o2_add_executable(
-        preclusters-reader-workflow
-        SOURCES src/preclusters-reader-workflow.cxx
-        COMPONENT_NAME mch
-        PUBLIC_LINK_LIBRARIES O2::MCHWorkflow)
 
 o2_add_executable(
         preclusters-to-clusters-original-workflow
